+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: `BackingCard`: Fix text input width.
+- Feature: Add `HorizontalCrowdfundingCard` molecule component.
+- Feature: Add `FlexWrapper` layout component.
 
 ## [4.0.2] - 2021-07-27